// src/models/option.model.ts
export interface Option {
  name?: string;
  points?: number;
  optionType: string;
  hp?: number | null;
  tenacity?: number | null;
  range?: number[] | null;
  damage?: string | null;
  reloading?: number | null;
  charge?: number | null;
  limited?: number | null;
  unique?: boolean;
  tags?: string[] | null;
  desc?: string | null;
  options?: Option[];
}